move UBX-MON-VER to mon folder
