+++ ublox_dgnss_node/include/ublox_dgnss_node/ubx/ubx_mon.hpp
@@ -21,65 +21,12 @@
 #include <string>
 #include <vector>
 #include "ublox_dgnss_node/ubx/ubx.hpp"
+#include "ublox_dgnss_node/ubx/mon/ubx_mon_ver.hpp"
 
 namespace ubx::mon
 {
 
-class MonVerPayload : UBXPayload
-{
-public:
-  static const msg_class_t MSG_CLASS = UBX_MON;
-  static const msg_id_t MSG_ID = UBX_MON_VER;
-
-  ch_t sw_version[30];        // null-terminated software version string
-  ch_t hw_version[10];        // null-terminated hardware version string
-  std::vector<std::string> extension;     // extended software information strings
-
-public:
-  MonVerPayload()
-  : UBXPayload(MSG_CLASS, MSG_ID)
-  {
-  }
-
-  MonVerPayload(ch_t * payload_polled, u2_t size)
-  : UBXPayload(MSG_CLASS, MSG_ID)
-  {
-    payload_.clear();
-    payload_.reserve(size);
-    payload_.resize(size);
-    memcpy(payload_.data(), payload_polled, size);
-
-    auto ptr = payload_.data();
-
-    memcpy(&sw_version, ptr, sizeof(sw_version) );
-    ptr += 30;
-    memcpy(&hw_version, ptr, sizeof(hw_version) );
-    ptr += 10;
-
-    while (ptr < (payload_.data() + size)) {
-      std::string s1(reinterpret_cast<char *>(ptr));
-      extension.push_back(s1);
-      ptr += 30;
-    }
-  }
-  std::tuple<u1_t *, size_t> make_poll_payload() override
-  {
-    payload_.clear();
-    return std::make_tuple(payload_.data(), payload_.size());
-  }
-  std::string to_string()
-  {
-    std::ostringstream oss;
-    oss << "sw_version: " << sw_version;
-    oss << " hw_version: " << hw_version;
-    for (auto e : extension) {
-      oss << " " << e;
-    }
-    return oss.str();
-  }
-};
-
-typedef UBXFrameComms<mon::MonVerPayload, usb::Connection> UbxMonVerFrameComms;
+typedef UBXFrameComms<mon::ver::MonVerPayload, usb::Connection> UbxMonVerFrameComms;
 
 class UbxMon
 {
@@ -96,6 +43,17 @@ public:
   std::shared_ptr<UbxMonVerFrameComms> ver()
   {
     return ver_;
+  }
+  void frame(std::shared_ptr<ubx::Frame> frame)
+  {
+    switch (frame->msg_id)
+    {
+      case UBX_MON_VER:
+        ver_->frame(frame);
+        break;
+      default:
+        break;
+    }
   }
 };
 }  // namespace ubx::mon
